Add Altera SoC Development Kit support
Use socfpga-3.15 and own dts. It is necessary to restrict
U-Boot when loading initrd, eg.

 > setenv initrd_high 0x01ffffff

Included u-boot script to be included on the SD card. If using
the original setup of the Altera SoC Kit board then it is
possible to just copy the u-boot.scr there and the U-Boot
process will call it automatically. The u-boot.scr is created
by the post-build script of Buildroot into the images/ directory.